+++ syft/source/directorysource/directory_source.go
@@ -4,8 +4,10 @@ import (
 	"fmt"
 	"os"
 	"path/filepath"
+	"strings"
 	"sync"
 
+	"github.com/bmatcuk/doublestar/v4"
 	"github.com/opencontainers/go-digest"
 
 	"github.com/anchore/syft/internal/log"
@@ -43,7 +45,7 @@ func New(cfg Config) (source.Source, error) {
 	}
 
 	if !fileMeta.IsDir() {
-		return nil, fmt.Errorf("given path is a file: %q", cfg.Path)
+		return nil, fmt.Errorf("given path is not a directory: %q", cfg.Path)
 	}
 
 	return &directorySource{
@@ -53,11 +55,11 @@ func New(cfg Config) (source.Source, error) {
 	}, nil
 }
 
-func (s *directorySource) ID() artifact.ID {
+func (s directorySource) ID() artifact.ID {
 	return s.id
 }
 
-func (s *directorySource) Describe() source.Description {
+func (s directorySource) Describe() source.Description {
 	name := cleanDirPath(s.config.Path, s.config.Base)
 	version := ""
 	supplier := ""
@@ -95,7 +97,7 @@ func (s *directorySource) FileResolver(_ source.Scope) (file.Resolver, error) {
 		return s.resolver, nil
 	}
 
-	exclusionFunctions, err := source.GetDirectoryExclusionFunctions(s.config.Path, s.config.Exclude.Paths)
+	exclusionFunctions, err := GetDirectoryExclusionFunctions(s.config.Path, s.config.Exclude.Paths)
 	if err != nil {
 		return nil, err
 	}
@@ -118,6 +120,60 @@ func (s *directorySource) Close() error {
 
 	s.resolver = nil
 	return nil
+}
+
+func GetDirectoryExclusionFunctions(root string, exclusions []string) ([]fileresolver.PathIndexVisitor, error) {
+	if len(exclusions) == 0 {
+		return nil, nil
+	}
+
+	// this is what directoryResolver.indexTree is doing to get the absolute path:
+	root, err := filepath.Abs(root)
+	if err != nil {
+		return nil, err
+	}
+
+	// this handles Windows file paths by converting them to C:/something/else format
+	root = filepath.ToSlash(root)
+
+	if !strings.HasSuffix(root, "/") {
+		root += "/"
+	}
+
+	var errors []string
+	for idx, exclusion := range exclusions {
+		// check exclusions for supported paths, these are all relative to the "scan root"
+		if strings.HasPrefix(exclusion, "./") || strings.HasPrefix(exclusion, "*/") || strings.HasPrefix(exclusion, "**/") {
+			exclusion = strings.TrimPrefix(exclusion, "./")
+			exclusions[idx] = root + exclusion
+		} else {
+			errors = append(errors, exclusion)
+		}
+	}
+
+	if errors != nil {
+		return nil, fmt.Errorf("invalid exclusion pattern(s): '%s' (must start with one of: './', '*/', or '**/')", strings.Join(errors, "', '"))
+	}
+
+	return []fileresolver.PathIndexVisitor{
+		func(_, path string, info os.FileInfo, _ error) error {
+			for _, exclusion := range exclusions {
+				// this is required to handle Windows filepaths
+				path = filepath.ToSlash(path)
+				matches, err := doublestar.Match(exclusion, path)
+				if err != nil {
+					return nil
+				}
+				if matches {
+					if info != nil && info.IsDir() {
+						return filepath.SkipDir
+					}
+					return fileresolver.ErrSkipPath
+				}
+			}
+			return nil
+		},
+	}, nil
 }
 
 // deriveIDFromDirectory generates an artifact ID from the given directory config. If an alias is provided, then
