+++ syft/source/exclude.go
@@ -1,69 +1,5 @@
 package source
 
-import (
-	"fmt"
-	"os"
-	"path/filepath"
-	"strings"
-
-	"github.com/anchore/syft/syft/internal/fileresolver"
-	"github.com/bmatcuk/doublestar/v4"
-)
-
 type ExcludeConfig struct {
 	Paths []string
-}
-
-func GetDirectoryExclusionFunctions(root string, exclusions []string) ([]fileresolver.PathIndexVisitor, error) {
-	if len(exclusions) == 0 {
-		return nil, nil
-	}
-
-	// this is what directoryResolver.indexTree is doing to get the absolute path:
-	root, err := filepath.Abs(root)
-	if err != nil {
-		return nil, err
-	}
-
-	// this handles Windows file paths by converting them to C:/something/else format
-	root = filepath.ToSlash(root)
-
-	if !strings.HasSuffix(root, "/") {
-		root += "/"
-	}
-
-	var errors []string
-	for idx, exclusion := range exclusions {
-		// check exclusions for supported paths, these are all relative to the "scan root"
-		if strings.HasPrefix(exclusion, "./") || strings.HasPrefix(exclusion, "*/") || strings.HasPrefix(exclusion, "**/") {
-			exclusion = strings.TrimPrefix(exclusion, "./")
-			exclusions[idx] = root + exclusion
-		} else {
-			errors = append(errors, exclusion)
-		}
-	}
-
-	if errors != nil {
-		return nil, fmt.Errorf("invalid exclusion pattern(s): '%s' (must start with one of: './', '*/', or '**/')", strings.Join(errors, "', '"))
-	}
-
-	return []fileresolver.PathIndexVisitor{
-		func(_, path string, info os.FileInfo, _ error) error {
-			for _, exclusion := range exclusions {
-				// this is required to handle Windows filepaths
-				path = filepath.ToSlash(path)
-				matches, err := doublestar.Match(exclusion, path)
-				if err != nil {
-					return nil
-				}
-				if matches {
-					if info != nil && info.IsDir() {
-						return filepath.SkipDir
-					}
-					return fileresolver.ErrSkipPath
-				}
-			}
-			return nil
-		},
-	}, nil
 }
